ai-service/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
python-dotenv==1.0.0
joblib==1.3.2
